+++ changelog.md
@@ -1,1 +1,1 @@
-- Add group chats
+- Improved positional audio